add definition of SMDP
Derive the simplified SMDP from a natural formulation and explain how
discounting should be handled in the Q-learning update. Furthermore, we
discuss the results from Hofri & Ross (1987) on the two-queue with
switc-over time. We propose an extension of their model by allowing the
server to idle for any arbitrary (finite) amount of time instead of
always idling indefinitely.