rewrite partial ordering, behavioral cloning
Rewrite sections on Partial Ordering, change the Proposition to be a
direct (without proof) Corollary of the "platoon preservation theorem".
Rewrite the introduction of the "lane automaton" to better fit the
introduction of the simple "threshold heuristic".
We removed the "factorization" of the conditional probability model.
We don't need this anymore, because we already introduced the state ->
action mapping.