fix neural embedding
We had `obs[n, l, :length]`, but that should have been `obs[n, l,
1:1+length]`, because the first entry indicates the length of the
horizon.

We now reverse the order of the horizon before feeding it to the RNN, in
order to make sure that vehicles which are crossing fist are processed
last and hence have the most influence on the embedding.